Test for lambda ingredients compiles, but checks!
Situation: it seems that the lambda ingredients are equal in both
versions (NISSA and our 'optimized' code). There are 2 remarks though:
1.The current code compiles, but there was a bracket missing in the
NISSA extracted part (mistake by me), so the results could be incorrect.
It is very unlikely, though, since the results obtained by the code
written by Francesco are identical.
2. Notice that b11 and b21 are almost identical! But possibly this is
not a problem.