GCC version has "homebrew" autotoools.
A makefile can be generated with the command

./generate_makefile.py
../src/{OpenAcc,RationalApprox,Include,Rand,Meas}/*.[ch] > makefile

inside the /build directory.
The main program compiles and executes. The results in gauge_meas.dat
 are not equal to TEST_REORDER/gauge_meas.dat, but the metropolis test
succeds.
Still to check if this can be compiled with pgcc withous problems (my
hope is that it's just necessary to modify the compiler and linker
variables in generate_makefile.py.