Added a bunch of test, most notably a total unitarity check on the device
Modified also the generate_dependence.py script to list also
the dependences on the standard libraries (this has not been tested yet)

The unitarity check just sums all deviations for all links, it would be
cool to have a maximum deviation too.

Now, on a random cold config this nuitarity check gives 1.926439e-314,
perhaps it is not true. More checks are in order.