Comparison with NISSA in progress. Tools added.
Added some "verbosity_lv" instruction in the code.
Added an "eval rational approx" tool in tools/.
Fixed some rational approxes taken from the conf in Nissa.
The situation is:
1. The larger amount of iterations in this code compared to Nissa
(see output of the "main" target) is due to the fact that the residue in
Nissa is a relative residue to the norm of the input vector. In our
code, the residue is instead absolute, and the norm of the vector is in
the order of 1e5, so the nissa version performs less iterations.
2. The error in the integration of the molecular dynamics equation
scales as expected like const/N**2 with N equal to no_md_steps (or a similar
variable), the total time being the same. Unfortunately, the 'const' is
1-2 orders of magnitude bigger than the one in Nissa.
Referring to point 1: by changing the algorithm to mimic Nissa (which
amounts to increase the residual by a factor of ~1e2) the integration
error increase of the order of 7% (2.29 vs 2.12). This strongly
indicates that the precision of the inversions and the precision of the
rational approximations is enough: if it were not so, the error ds would
not go to zero by increasing N.

3. When instead the gauge substeps are increased from 5 to 10, ds goes
to 0.33. There could be some problems in the Omelyan integrator?
I will read the two codes and compare/