Written (not tested) input file mechanism.
Code compiles but, since the input file is missing, of course cannot be
tested.
In order to generate the input file, it makes sense to write a routine
that will create a template if the file is not found.
TODO: it seems reasonable to move the names of the parameters to the
place where they are defined now (inside the structures that contain
them). In practice, all the strings containing parameter names and
names of goups of parameters should be moved from Include/init.c
in the files included.