Improved generate makefile, rationalizations, resuscitated deodoe test tool
The generate_makefile.py script has been improved, now it is possible to generate
a makefile with, for example, the command

./generate_makefile.py ../src/OpenAcc/main.c ../src/DbgTools/deo_doe_test.c

and obtaining a makefile with 2 (final) make targets, that is 'main' and
'deo_doe_test'.

The deo_doe_test has also been resuscitated and improved
(it will be really necessary in the future).