Found some issues, the staggered phases of course.
The staggere phases are saved with the matrices, bu unfortunately
only the first two rows are multiplied for the phase.
This makes the staggered phase undetectable, and comparison between
'staggered-less' code and 'staggered-full' code more complicated,
in order to make this comparison it will be necessary to change the
old code (this is what the branch 'StagPhasesRewrite_CHECK' is
here for).