All seems to work (metropolis test succeeds on a 16^4)
Cause of the issue:

instead of

which caused the force to be zero.
Other things:
Some 4D cycles with 4 nested loops could be 1D-ized.
(it has been done in a case).
Force calculation routine has been implemented (it's just a copy of the
routine to calculate the momenta part of the action).
notice: even if the fermion part of the force seems to be quite smaller
than the gauge part, it seems that the gauge substeps needed are not so
many (that is, accuracy in the fermionic part is more crucial than a
naive analysis would suggest? Food for tought).