Here single cpu works with stouting.
Modification to make NORANDOM compliant the initialization of the
random number generator.
I know it does not make sense, but there are cases in which the
NORANDOM mode is not completely deterministic; in particular it seems
that there are differences in the calculation of the final action, in
the calculation of the eigenvalues.