Now it accepts even with stout_steps = 0
In the fermion force calculation, the calculation
of sigma from sigma prime was performed even when stout_steps = 0,
which is wrong.
The reason why it was not accepting even  with stouting (in the old
settings) was a number of gauge substeps too low (1).
This might mean that the code is not correct in the case of only 1
gauge substep?