Corrected Re(*)=0 problem in fermionic measures.
It was due to the lack of a reduction pragma.
It is quite strange though that there was a reduction(+:resR)
and the missing pragma was instead reduction(+:resI), but it was the
REAL part to be zero.
It seems a bug of the compiler (pgi15.9), possibly confused by the fact that
resR and resI were both reduced over, and this lead it to choose the
wrong one (??).

Notice that the bugged function was not used during molecular dynamics
evolution, but just in fermionic measurements.

Still to check.