Single precision MD seems to work but it's slower than double prec
A mistake/misunderstanding was corrected, the conversion functions only
worked on the first element of the usual 8-vector for the gauge conf
and the momenta.

After this correction, the algorithm accepts, but it's very slow.

Now, checking issues in the code conversion tool to see what went wrong
(also, comparing with an old branch were single precision multishift
inversion was already working, and faster).

Moreover: corrected setting_file_parser.c to have a complete template
written.