IMPORTANT: Lat dims must be passed as arguments at compile time.
The 'lattice_dimensions.h' file does not exist anymore;
this allows to have ONLY one directory with the code and an arbitrary
build location.

In fact, the 'lattice_dimensions.h' file was included in
'src/Mpi/geometry_multidev.h' as

\#include "../../build/lattice_dimensions.h"

And that forced a non-intelligent way of working.
Now, it should be sufficient to copy the makefile in any directory in
the filesystem, and any 'make' should work.

EXCEPT: 'make all' and 'make' don't work because they also try to
        compile 'rgen', which is not where the makefile expects it to be
        found.