An error in the 'source_norm' calculation corrected.
Found the issue which prevented the acceleration in
fermion force calculation:

The 'source_norm' was calculated as the norm of the starting residue
and not as the norm of the input vector.

Since the stop condition was delta/source_norm < res^2, this
forced the single inversion post multi inversion to do still many steps,
to reduce the residue much lower beyond the intended target.