On multi gpu, Measures fail if MD is performed before.
Implemented some diagnostics (= put some printfs in the code).

Also, found a couple of instructions which, if removed, would
remove the issue (2 functions used in the stouting machinery,
in the sigma_p to sigma function).
This does not make sense.

BUT: in general, this branch of the code has NOT been tested
in multinode since LONG ago. Plan: go back in time and change
the fate.