Corrected reason for wrong behaviour in ferm measurements
The issue was lying in the non initialization of the trial solution
in the measurements.
The issue was present only when multi gpu was used, when measures and
update were performed (which made the issue slightly difficult to
debug). Of course, this situation is the most common.
In fact, it is not clear to me why the bug would appear only in this
situation.
It may be important to check wether this issue can be present in other
situations.