Implemented (hopefully) solution for the lack of convergence
Bug: some cases, very rare, it happens that convergence with mixed
precision is not reached (but res/res_target= 1.00000 is displayed)
This may be due to the infrequency of the double precision steps.
However, a "SAFETY_MARGIN" define has been added which forces the
inverter to go a bit further: hopefully this will fix everything
(otherwise it may be clever to increase the frequency of double
precision steps).