Benchmark tools restructured, other improvements
(HUGE) Modifications to benchmark makefile.
To do: write guide!

Moreover:
Improvements to double-single precision converters
(only file timestamps are now checked in order to decide
if a single precision listing has to be rewritten, but the
old thorough checks are available with the 'recheck' option.

Added time measurements of gauge md routines in md diagnostics file.