Made makefile generation more versatile
Now there is no need to set the compiler before creating the makefile.
The compiler is choosen only at compilation time, with

make MODE=GNU # or PGI,PGISLOW,MPI...

the settings and the logic for the compiler and the flags are written in the file

   compiler_setting_library.txt

which is copied totally in the produced makefile.