Supposedly fixed hideous bug where simple read/write fail with PGI.
When compiling with PGI, the content of the program_status.txt
file was not written correctly in the appropriate part of
the mc_params structure, resulting in strange numbers with no meaning
being read.
This has possibly to do with some misused '#define' or '#include'.
By now, the problem should be solved.