Added rational approx generation to autotools (with tricks)
Problem: rationalapprox.c must be compiled with c and c++ compiler
depending on the final target (c in the main program, c++ only for
rational approx generator).

I moved all the sources for rational approx generation to 'tools',
created a symlink to rationalapprox.c called 'rationalapprox.cpp', which
gets compiled with g++ as all the sources for rational approx
generation.